+++ api/src/main/java/ee/tlu/evkk/api/controller/ExerciseController.java
@@ -1,0 +1,32 @@
+package ee.tlu.evkk.api.controller;
+
+import ee.tlu.evkk.api.service.ExerciseService;
+import ee.tlu.evkk.dal.dto.Exercise;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/exercises")
+@RequiredArgsConstructor
+@CrossOrigin("*")
+public class ExerciseController {
+
+  private final ExerciseService exerciseService;
+
+  @GetMapping
+  public List<Exercise> getAll() {
+    return exerciseService.getAllExercises();
+  }
+
+  @GetMapping("/{id}")
+  public Exercise getById(@PathVariable Long id) {
+    return exerciseService.getExerciseById(id);
+  }
+
+  @PostMapping
+  public void insert(@RequestBody Exercise exercise) {
+    exerciseService.insertExercise(exercise);
+  }
+}
