+++ dal/src/main/java/ee/tlu/evkk/dal/dto/Category.java
@@ -1,0 +1,19 @@
+package ee.tlu.evkk.dal.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.io.Serializable;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Category implements Serializable {
+  private Long id;
+  private String name;
+}
