+++ dal/src/main/resources/ee/tlu/evkk/dal/dao/LanguageLevelDao.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="ee.tlu.evkk.dal.dao.LanguageLevelDao">
+
+  <select id="findAllLanguageLevels" resultType="ee.tlu.evkk.dal.dto.LanguageLevel">
+    SELECT id, level FROM core.language_level ORDER BY id
+  </select>
+
+</mapper>
