--- ui/src/elle/components/library/SearchResult.js
@@ -1,4 +1,0 @@
-//siia tuleb tulemuste komponent
-import React from 'react';
-import { Link } from 'react-router-dom';
-
