--- ui/src/elle/components/library/h5p/storage/README.txt
@@ -1,1 +1,0 @@
-siin kaustas hoiab ajutiselt H5P JSON-e kuniks meeskond leiab parema viisi hoiustada neid
