+++ ui/src/elle/components/library/json/categories.json
@@ -12,4 +12,3 @@
   { "id": 11, "name": "Sõnaliigid" },
   { "id": 12, "name": "Kuulamine ja arusaamine" }
 ]
-
