--- ui/src/elle/components/library/json/time.json
@@ -1,7 +1,0 @@
-[
-  { "id": 1, "value": 5, "label": "Kuni 5"},
-  { "id": 2, "value": 10, "label": "Kuni 10"},
-  { "id": 3, "value": 15, "label": "Kuni 15"},
-  { "id": 4, "value": 20, "label": "Kuni 20"},
-  { "id": 5, "value": 25, "label": "Kuni 25"}
-]
