+++ ui/src/elle/components/library/search/SearchBar.js
@@ -1,0 +1,15 @@
+import { TextField, Button } from '@mui/material'
+
+
+export default function SearchBar() {
+    return (
+        <div className="library-search-box">
+            <TextField label="Otsing"
+                variant="outlined"
+                name="otsing"
+                style={{ width: '100%', maxWidth: '50%' }}
+            ></TextField>
+            <Button>Otsi</Button>
+        </div>
+    )
+}
