+++ ui/src/elle/components/library/search/SearchResult.js
@@ -1,0 +1,4 @@
+//siia tuleb tulemuste komponent
+import React from 'react';
+import { Link } from 'react-router-dom';
+
